Clean up some stuff
